Rename scripts.util to scripts.library
