+++ .ruff.toml
@@ -29,7 +29,6 @@ lint.extend-ignore = [
   "G004",   # Logging statement uses f-string
   "PLR",    # Design related pylint codes
   "E501",   # Line too long
-  "PT004",  # Use underscore for non-returning fixture (use usefixture instead)
   "RET505", # Unnecessary `else` after `return` statement
   "E402", # Module level imports at top of file need disable for numba default imports
   "NPY201", # Numpy 2 warnings ignore for the moment
