+++ rules/ann.smk
@@ -4,7 +4,7 @@ to apply the ann and risetime cuts for psd.
 
 """
 
-from scripts.util.patterns import (
+from scripts.library.patterns import (
     get_pattern_tier,
     get_pattern_log,
     get_pattern_pars,
