+++ rules/blinding_calibration.smk
@@ -4,7 +4,7 @@ Snakemake rules for calibrating daq energy for blinding. Two steps:
 - combining all channels into single par file
 """
 
-from scripts.util.patterns import (
+from scripts.library.patterns import (
     get_pattern_pars,
     get_pattern_plts,
     get_pattern_pars_tmp_channel,
