+++ rules/main.smk
@@ -1,6 +1,6 @@
 import os
 from datetime import datetime
-from scripts.util.utils import (
+from scripts.library.utils import (
     filelist_path,
     log_path,
     tmp_par_path,
