+++ rules/qc_phy.smk
@@ -1,7 +1,7 @@
-from scripts.util.pars_loading import ParsCatalog
-from scripts.util.create_pars_keylist import ParsKeyResolve
-from scripts.util.utils import filelist_path, set_last_rule_name
-from scripts.util.patterns import (
+from scripts.library.pars_loading import ParsCatalog
+from scripts.library.create_pars_keylist import ParsKeyResolve
+from scripts.library.utils import filelist_path, set_last_rule_name
+from scripts.library.patterns import (
     get_pattern_pars_tmp_channel,
     get_pattern_plts_tmp_channel,
     get_pattern_log_channel,
