+++ scripts/build_dsp.py
@@ -9,7 +9,7 @@ from dspeed import build_dsp
 from legendmeta import LegendMetadata, TextDB
 from legendmeta.catalog import Props
 from lgdo import lh5
-from util.log import build_log
+from library.log import build_log
 
 
 def replace_list_with_array(dic):
