+++ scripts/build_skm.py
@@ -5,7 +5,7 @@ from legendmeta import TextDB
 from legendmeta.catalog import Props
 from lgdo import lh5
 from lgdo.types import Array, Struct, Table, VectorOfVectors
-from util.log import build_log
+from library.log import build_log
 
 
 def get_all_out_fields(input_table, out_fields, current_field=""):
