+++ scripts/library/patterns.py
