+++ scripts/pars_dsp_dplms.py
@@ -9,8 +9,8 @@ import numpy as np
 from legendmeta import LegendMetadata, TextDB
 from legendmeta.catalog import Props
 from lgdo import Array, Table
+from library.log import build_log
 from pygama.pargen.dplms_ge_dict import dplms_ge_dict
-from util.log import build_log
 
 argparser = argparse.ArgumentParser()
 argparser.add_argument("--fft_raw_filelist", help="fft_raw_filelist", type=str)
