+++ scripts/pars_dsp_nopt.py
@@ -8,9 +8,9 @@ import numpy as np
 import pygama.pargen.noise_optimization as pno
 from legendmeta import LegendMetadata, TextDB
 from legendmeta.catalog import Props
+from library.log import build_log
 from pygama.pargen.data_cleaning import generate_cuts, get_cut_indexes
 from pygama.pargen.dsp_optimize import run_one_dsp
-from util.log import build_log
 
 sto = lh5.LH5Store()
