+++ scripts/pars_hit_qc.py
@@ -12,14 +12,14 @@ import numpy as np
 from legendmeta import LegendMetadata, TextDB
 from legendmeta.catalog import Props
 from lgdo.lh5 import ls
+from library.convert_np import convert_dict_np_to_float
+from library.log import build_log
 from pygama.pargen.data_cleaning import (
     generate_cut_classifiers,
     get_keys,
     get_tcm_pulser_ids,
 )
 from pygama.pargen.utils import load_data
-from util.convert_np import convert_dict_np_to_float
-from util.log import build_log
 
 log = logging.getLogger(__name__)
