+++ scripts/pars_tcm_pulser.py
@@ -6,8 +6,8 @@ import lgdo.lh5 as lh5
 import numpy as np
 from legendmeta import LegendMetadata, TextDB
 from legendmeta.catalog import Props
+from library.log import build_log
 from pygama.pargen.data_cleaning import get_tcm_pulser_ids
-from util.log import build_log
 
 argparser = argparse.ArgumentParser()
 argparser.add_argument("--configs", help="configs path", type=str, required=True)
