+++ tests/test_util.py
@@ -1,7 +1,7 @@
 import json
 from pathlib import Path
 
-from scripts.util import (
+from scripts.library import (
     CalibCatalog,
     FileKey,
     pars_catalog,
@@ -9,8 +9,8 @@ from scripts.util import (
     subst_vars,
     unix_time,
 )
-from scripts.util.patterns import get_pattern_tier_daq, get_pattern_tier_dsp
-from scripts.util.utils import (
+from scripts.library.patterns import get_pattern_tier_daq, get_pattern_tier_dsp
+from scripts.library.utils import (
     par_dsp_path,
     par_overwrite_path,
     tier_dsp_path,
